Moved home screen into separate files
